+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Maybe:
 ## 3.6.7
 
 - rename doe, write_does and load_does to `sweep` module `read_sweep`, `write_sweep` ...
+- Route and Routes are pydantic.BaseModel instead of dataclasses
 
 ## 3.6.6